backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
sentence-transformers==3.3.1
numpy>=1.24.0
pydantic==2.10.4
python-multipart==0.0.20
